fix code placement
